% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fun_sef}
\alias{fun_sef}
\title{Calculate Standardized Difference between two groups.(category)}
\usage{
fun_sef(input_data, treat_vec_name)
}
\arguments{
\item{input_data}{input dataframe with 2 columns(treatmend and responce)}

\item{treat_vec_name}{treatment columns name}
}
\value{
standardized differnce
}
\description{
Calculate Standardized Difference between two groups.(category)
}

