% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fun_sef_cat}
\alias{fun_sef_cat}
\title{Calculate Standardized Difference between two groups.(category)}
\usage{
fun_sef_cat(vec_tr, vec_con)
}
\arguments{
\item{vec_tr}{treatment group category vector}

\item{vec_con}{control group category vector}
}
\value{
standardized differnce
}
\description{
Calculate Standardized Difference between two groups.(category)
}

