% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fun_sef_df}
\alias{fun_sef_df}
\title{Calculate Standardized Difference between two groups with dataframe.}
\usage{
fun_sef_df(input_data, treat_vec_name)
}
\arguments{
\item{input_data}{input dataframe contain treatment variable and responce variables}

\item{treat_vec_name}{treatment columns name}
}
\value{
standardized differnce vector
}
\description{
Calculate Standardized Difference between two groups with dataframe.
}

