% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fun_sef_num}
\alias{fun_sef_num}
\title{Calculate Standardized Difference between two groups.(numeric)}
\usage{
fun_sef_num(vec_tr, vec_con)
}
\arguments{
\item{vec_tr}{treatment group numeric vector}

\item{vec_con}{control group numeric vector}
}
\value{
standardized differnce
}
\description{
Calculate Standardized Difference between two groups.(numeric)
}

